requirements.txt
Django==1.11.6
django-tenant-schemas==1.9.0
graphene==2.0
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
iso8601==0.1.12
pkg-resources==0.0.0
promise==2.1
psycopg2==2.7.3.2
pytz==2017.2
Rx==1.6.0
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.2
